% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacama.waxes-package.R
\docType{package}
\name{atacama.waxes-package}
\alias{atacama.waxes}
\alias{atacama.waxes-package}
\title{atacama.waxes: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mat1506/atacama.waxes}
  \item Report bugs at \url{https://github.com/mat1506/atacama.waxes/issues}
}

}
\author{
\strong{Maintainer}: Matías Frugone-Álvarez \email{matutefrugone@gmail.com} (\href{https://orcid.org/0000-0002-4182-9902}{ORCID}) [copyright holder]

}
\keyword{internal}
