% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacama.waxes-package.R
\docType{package}
\name{atacama.waxes-package}
\alias{atacama.waxes}
\alias{atacama.waxes-package}
\title{atacama.waxes: Hydroclimate variations over the last 17,000 years as measured by leaf waxes in landscape physiognomy and rodent middens from the south-central Atacama Desert, Chile}
\description{
This package is a research compendium of the manuscript "Hydroclimate variations over the last 17,000 years as measured by leaf waxes in rodent middens from the south-central Atacama Desert, Chile", it constitutes a collection of all the digital parts of a research project, including data, code, texts (protocols, reports, questionnaires, metadata).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mat1506/atacama.waxes}
  \item Report bugs at \url{https://github.com/mat1506/atacama.waxes/issues}
}

}
\author{
\strong{Maintainer}: Matías Frugone-Álvarez \email{matutefrugone@gmail.com} (\href{https://orcid.org/0000-0002-4182-9902}{ORCID}) [copyright holder]

}
\keyword{internal}
